Restructure server into smaller submodules
